feat(#19): Created a data.json file and adapted the code
Created the file in the directory /map/data/data.json. Adapted the code in Map.tsx and Sidebar.tsx to work with the data from this file. This provides a better and cleaner overview of both the code and the data, which will help me improve the map UI. I will add a new data schema in the next step.